refactor: InputView.java -> Input
