+++ src/main/java/rentCompany/RentCompanyController.java
@@ -1,7 +1,7 @@
 package rentCompany;
 
 import rentCompany.model.*;
-import rentCompany.view.InputView;
+import rentCompany.view.Input;
 import rentCompany.view.Output;
 
 public class RentCompanyController {
@@ -9,11 +9,11 @@ public class RentCompanyController {
     public static void main(String[] args) {
         Company company = CompanyFactory.createCompany();
 
-        company.addCar(new K5(InputView.inputDistance()));
-        company.addCar(new Sonata(InputView.inputDistance()));
-        company.addCar(new Sonata(InputView.inputDistance()));
-        company.addCar(new Avante(InputView.inputDistance()));
-        company.addCar(new Avante(InputView.inputDistance()));
+        company.addCar(new K5(Input.inputDistance()));
+        company.addCar(new Sonata(Input.inputDistance()));
+        company.addCar(new Sonata(Input.inputDistance()));
+        company.addCar(new Avante(Input.inputDistance()));
+        company.addCar(new Avante(Input.inputDistance()));
 
         Output.printReport(company.makeReport());
     }
